refactor: move sdk init code
