+++ embrace-android-sdk/src/test/java/io/embrace/android/embracesdk/internal/api/delegate/LogsApiDelegateTest.kt
@@ -28,7 +28,7 @@ internal class LogsApiDelegateTest {
         val moduleInitBootstrapper = fakeModuleInitBootstrapper(
             logModuleSupplier = { _, _, _, _, _, _, _ -> FakeLogModule(logService = logService) }
         )
-        moduleInitBootstrapper.init(ApplicationProvider.getApplicationContext(), 0)
+        moduleInitBootstrapper.init(ApplicationProvider.getApplicationContext())
         orchestrator = moduleInitBootstrapper.sessionOrchestrationModule.sessionOrchestrator as FakeSessionOrchestrator
 
         val sdkCallChecker = SdkCallChecker(FakeEmbLogger(), FakeTelemetryService())
